% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CgdsSNPpairMatrix}
\alias{CgdsSNPpairMatrix}
\title{Get's SNP-block pair}
\usage{
CgdsSNPpairMatrix(filteredsnp)
}
\arguments{
\item{filename, }{string with route to file with .vcf, .bed or .gds data type}
}
\value{
converted gds filename
}
\description{
Get's SNP-block pair
}
