% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CgetCorrRecombClust.R
\name{CgetCorrRecombClust}
\alias{CgetCorrRecombClust}
\title{Recombination correlation matrix}
\usage{
CgetCorrRecombClust(probmat, annot, window)
}
\arguments{
\item{annot}{Matrix with initial and final coordinades of the model.}

\item{window}{window size.}

\item{mat}{Matrix with model probabilities - recombClust output.}

\item{range}{range where the function is executed.}
}
\value{
A vector with chromosome probabilities to be in recomb population
}
\description{
Get correlation matrix from the probability of chromosome recombination
}
