% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CGetDatafromFile}
\alias{CGetDatafromFile}
\title{Covnert input data file, .vcf or .bed to gds}
\usage{
CGetDatafromFile(file)
}
\arguments{
\item{filename, }{string with route to file with .vcf, .bed or .gds data type}
}
\value{
converted gds filename
}
\description{
Covnert input data file, .vcf or .bed to gds
}
