% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CgetIndexfromGenCoord.R
\name{CgetIndexfromGenCoord}
\alias{CgetIndexfromGenCoord}
\title{This function get index position from genomic coordinates}
\usage{
CgetIndexfromGenCoord(
  filename,
  chrom,
  gcstart = NULL,
  gcend = NULL,
  minmaf = 0.1
)
}
\arguments{
\item{filename}{character string with file path}

\item{chrom}{chromosome}

\item{gcstart}{genomic coordinates - start}

\item{gcend}{genomic coordinates - end}

\item{minmaf}{minimum MAF, default 0.01}
}
\value{
data ready to be processed
}
\description{
This function get index position from genomic coordinates
}
