% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cGetRecombProb}
\alias{cGetRecombProb}
\title{Get the probability of chromosome recombination}
\usage{
cGetRecombProb(probmat, annot, window)
}
\arguments{
\item{annot}{Matrix with initial and final coordinades of the model.}

\item{window}{window size.}

\item{mat}{Matrix with model probabilities - recombClust output.}

\item{range}{range where the function is executed.}
}
\value{
A vector with chromosome probabilities to be in recomb population
}
\description{
Get the probability of chromosome recombination
}
