% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CLQD_mod}
\alias{CLQD_mod}
\title{Estimation of LD block regions}
\usage{
CLQD_mod(
  CorMat,
  CLQcut = 0.2,
  clstgap = 40000L,
  hrstType = "fast",
  hrstParam = 200L,
  CLQmode = "maximal"
)
}
\arguments{
\item{CLQcut}{A numeric value of threshold for the correlation value |r|, between 0 to 1.}

\item{clstgap}{An integer value to specifying the threshold of physical distance (bp) between two consecutive SNPs
which do not belong to the same clique, i.e., if a physical distance between two consecutive SNPs in a clique
greater than \code{clstgap}, then the algorithm split the cliques satisfying each
clique do not contain such consecutive SNPs}

\item{hrstType, }{string value with hrstType=c("near-nonhrst", "fast", "nonhrst")}

\item{CLQmode}{A character string to specify the way to give priority among detected cliques, the algorithm gives priority to the largest clique

<output>}

\item{OCM}{Correlation matrix}

\item{leng}{An integer value to specify the number of SNPs in a preceding and a following region
of each sub-region boundary, every SNP in a preceding and every SNP in a following region need to be in weak LD.}

\item{MAFcut}{An numeric value to specifying the MAF threshold.}

\item{subSegmSize}{An integer value to specify the upper bound of the number of SNPs in a one-take sub-region.}

\item{appendRare}{If \code{appendRare = TRUE}, the algorithm append rare SNPs (MAF<MAFcut) to the constructed LD blocks or add a new LD blocks}

\item{checkLargest}{If \code{checkLargest = TRUE}, the algorithm use heuristic procedure to reduce runtime of CLQ-D execution}
}
\value{
A data frame of block estimation result.
Each row of data frame shows the starting SNP and end SNP of each estimated LD block.
}
\description{
\code{Big_LD} returns the estimation of LD block regions of given data.
}
