% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRecombProbabilities.R
\name{computeRecombProbabilities}
\alias{computeRecombProbabilities}
\title{Main function to run all the recombClust pipeline}
\usage{
computeRecombProbabilities(haplos, annot, range, window = 500, ...)
}
\arguments{
\item{annot}{GenomicRanges with the SNPs annotation}

\item{range}{Region GenomicRanges}

\item{window}{only to debug, by default window = 500bp}

\item{...}{Further arguments passed to runLDmixtureModel}

\item{filename}{File with genotypic data (SNPs and samples) the data file type must be .vcf (Variant Call Format), 
.gds (Genomic Data Structure) or .bed (Browser Extensible Data).}
}
\value{
A list with four elements:
\itemize{
 \item{porbTab: Summarized probabilities by window}
 \item{models: List of models}
}
}
\description{
This function takes as input phased chromosomes and returns the responsibilities
PCA and the cluster classification of the samples.
}
