% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRecombProbabilities_inFile.R
\name{computeRecombProbabilities_inFile}
\alias{computeRecombProbabilities_inFile}
\title{Main function to run all the recombClust pipeline with data in a file}
\usage{
computeRecombProbabilities_inFile(
  filename,
  range,
  samples = NULL,
  window = 500,
  minmaf = 0.1,
  resfilename = "recombClustResults.hdf5",
  resgroup = "AllModels",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{filename}{File with genotypic data (SNPs and samples) the data file 
type must be .vcf (Variant Call Format), .gds (Genomic Data Structure) or 
.bed (Browser Extensible Data).}

\item{range}{Region GenomicRanges}

\item{samples}{optional, character vector with sample ids to be used to compute models}

\item{window}{only to debug, by default window = 500bp}

\item{minmaf}{optional, threshold to filter by minimum MAF, by default minmaf = 0.1}

\item{resfilename}{string, path and file name to store results. By default
this runLDmixture stores data in recombClustResults.hdf5, if resfilename 
is empty then function works on memory. To work with big regions it is recommended 
to work with data on disk. Only use resfilename = "" when regions are small in order to 
prevent memory overflows and low performance.}

\item{resgroup}{string, folder inside a file where we want to store results. By default
runLDmixture stores data inside group "AllModels" in resfilename file.}

\item{overwrite}{boolean, (optional) either a logical value indicating whether
the output file can be overwritten or not, by default files are not overwritten.}

\item{...}{Further arguments passed to runLDmixtureModel}

\item{annot}{GenomicRanges with the SNPs annotation}
}
\value{
probTab: Summarized probabilities by window
}
\description{
This function takes as input phased chromosomes and returns the responsibilities
PCA and the cluster classification of the samples.
}
