% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_HDF_HiCBricks_File}
\alias{Create_HDF_HiCBricks_File}
\title{Write data to hdf5 file with HiCBricks structure}
\usage{
Create_HDF_HiCBricks_File(filename, mat, bintable)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{mat}{numerical mattrix with correlation mattrix}

\item{bintable}{dataframe with chromosome name
\itemize{
 \item{"chrom"}{character, chromosome name}
 \item{"start"}{numerical, range start position}
 \item{"end"}{numerical, range end position}
}}
}
\value{
none
}
\description{
Creates a hdf5 file with HiCBrick structure with mandatory information,
  Base.matrices (group)
     Chromosome (group)
  Base ranges (group)
     Bintable (group)
}
\details{
and the corresponding information.
}
