% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEmptyHdf5File}
\alias{createEmptyHdf5File}
\title{Create File}
\usage{
createEmptyHdf5File(filename, force = FALSE)
}
\arguments{
\item{filename}{string, path and file name to be created.}

\item{force}{boolean, if true this function overwrites existing group.}
}
\value{
integer, if 0 the process was successful and group was created in 
the file
}
\description{
Create an empty Hdf5 data file
}
