% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecombProb.R.Carlos.R
\name{createMatrix}
\alias{createMatrix}
\title{recombRes: resultado de CrunLDmixtureModel
range: GRanges de la región}
\usage{
createMatrix(recombRes, range)
}
\description{
recombRes: resultado de CrunLDmixtureModel
range: GRanges de la región
}
