% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMatrixfromModels.R
\name{createMatrixfromModels}
\alias{createMatrixfromModels}
\title{Get matrix from model}
\usage{
createMatrixfromModels(recombRes, range)
}
\arguments{
\item{recombRes}{Results from LDMixture}

\item{range}{Range to look in}

\item{window}{windows size, by default 500bp}
}
\value{
Probability matrix by window
}
\description{
Get matrix from models obtained by LDMixture
}
