% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CTransformtoSampleAlleles}
\alias{CTransformtoSampleAlleles}
\title{Convert 3D-Matrix to 2D-Matrix}
\usage{
CTransformtoSampleAlleles(x, allele, sample, variable)
}
\arguments{
\item{3D-matrix}{with allele, samples and SNPs information [allele,samples,SNPs]}
}
\value{
2D-matrix with one row with each sample-allel [sample-allele, SNP]
}
\description{
Convert 3D-Matrix to 2D-Matrix
}
