% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CWriteResults}
\alias{CWriteResults}
\title{Write data to file}
\usage{
CWriteResults(columns, results, outputfile)
}
\arguments{
\item{columns}{string vector with column names}

\item{results}{Numerical vector with data}

\item{outputfile}{relative path to output data file}
}
\value{
none
}
\description{
Write data to file
}
