% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{existsHdf5Element}
\alias{existsHdf5Element}
\title{Test if element exits}
\usage{
existsHdf5Element(filename, element)
}
\arguments{
\item{filename}{string, path and file name to search element.}

\item{element}{string, group or dataset name to search inside hdf5 data file.}
}
\value{
A boolean indicating whether or not the element being searched exists
}
\description{
Test if group name or dataset name exists inside the hdf5 data file
}
