% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_graph_matrix_data}
\alias{get_graph_matrix_data}
\title{Get graphs data from addjacency matrix}
\usage{
get_graph_matrix_data(OCM, hrstParam)
}
\arguments{
\item{OCM}{Correlation matrix}

\item{hrstParam}{A numeric value of threshold for the correlation value |r|, between 0 to 1.

<output>}
}
\value{
A list with , cores, highcore and local_cores
\itemize{
 \item{"adjMatrix"}{adjacent matrix graph}
 \item{"cores"}{a maximal subgraph in which each vertex has at least degree k. }
 \item{"highcore"}{Number of cores with "cores">hrstParam}
 \item{"local_cores"}{local_cores}
}
}
\description{
\code{Big_LD} returns adjacency matrix graph data to test graph complexity
}
