% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{This function takes a file as input, converts file to gds format
and get data ready to be processed}
\usage{
getData(filename, chromosome, range, minmaf = 0.1)
}
\arguments{
\item{filename}{character string with file path}

\item{chromosome}{character or character vector with chromosomes}

\item{range}{genomic ranges with regions to be read}

\item{minmaf}{minimum MAF, default 0.01}
}
\value{
data ready to be processed
}
\description{
This function takes a file as input, converts file to gds format
and get data ready to be processed
}
