% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getProbs}
\alias{getProbs}
\title{Compute cluster Recomb freq by mean of voting}
\usage{
getProbs(mat, sel)
}
\arguments{
\item{mat}{with initial probabilities}

\item{sel}{overlap ranges}
}
\value{
NumericVector with recombination probabilites
}
\description{
Compute cluster Recomb freq by mean of voting
}
