% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getProbs_hdf5}
\alias{getProbs_hdf5}
\title{Compute cluster Recomb freq by mean of voting}
\usage{
getProbs_hdf5(
  filename,
  group,
  dataset,
  selection,
  nCols,
  outgroup = NULL,
  outdataset = NULL
)
}
\arguments{
\item{filename}{with initial probabilities}

\item{group}{String, inside we have one dataset for each chromosome}

\item{dataset}{String, chromosome dataset inside the group}

\item{selection}{overlap ranges}

\item{nCols}{Integer, number of columns in selection.}

\item{outgroup}{String output group}

\item{outdataset}{String output dataset}
}
\value{
NumericVector with recombination probabilites
}
\description{
Compute cluster Recomb freq by mean of voting
}
