% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecombProb.R
\name{getRecombProb}
\alias{getRecombProb}
\title{Get matrix probabilities by window size}
\usage{
getRecombProb(mat, annot, range, window = 500)
}
\arguments{
\item{mat}{Results obtained with the different models LDMixture}

\item{annot}{Annotation obtained for the different models in LDMixture}

\item{range}{Range to look in}

\item{window}{windows size, by default 500bp}
}
\value{
Probability matrix by window
}
\description{
Get matrix probabilities by window size
}
