% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecombProb.R.Carlos.R
\name{getRecombProb1}
\alias{getRecombProb1}
\title{mat, annot: elementos de lista de createMatrix
range: ídem a range de createMatrix
window: tamaño de la ventana (500bp por defecto)}
\usage{
getRecombProb1(mat, annot, range, window)
}
\description{
mat, annot: elementos de lista de createMatrix
range: ídem a range de createMatrix
window: tamaño de la ventana (500bp por defecto)
}
