% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecombProb_hdf5.R
\name{getRecombProb_hdf5}
\alias{getRecombProb_hdf5}
\title{Get matrix probabilities by window size}
\usage{
getRecombProb_hdf5(filename, group, range, window = 500)
}
\arguments{
\item{filename}{Results obtained with the different models LDMixture}

\item{group}{Results obtained with the different models LDMixture}

\item{range}{Range to look in}

\item{window}{windows size, by default 500bp}

\item{annot}{Annotation obtained for the different models in LDMixture (A ELIMINAR !!)}
}
\value{
Probability matrix by window
}
\description{
Get matrix probabilities by window size
}
