% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindows.R
\name{getWindows}
\alias{getWindows}
\title{Get the windows within range}
\usage{
getWindows(annot, range, window = 500)
}
\arguments{
\item{annot}{GenomicRanges with the SNPs annotation}

\item{range}{Range to look in}

\item{window}{windows size, by default 500bp}
}
\value{
windows that overlap with the annotations within the range range
\itemize{
 \item{"chunk"}{Log-likelihood of mixture model}
 \item{"overLaps"}{Log-likelihood of linkage model}
}
}
\description{
Get the windows within range
}
