% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkageFreq.R
\name{linkageFreq}
\alias{linkageFreq}
\title{Reestimate linkage block frequencies}
\usage{
linkageFreq(Resp, Block)
}
\arguments{
\item{Resp}{Numerical with the responsibilities for linkage population}

\item{Block}{Character with the blocks genotypes}
}
\value{
Numerical with the new block frequencies for linkage population
}
\description{
Reestimate linkage block frequencies
}
