% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombFreq.R
\name{recombFreq}
\alias{recombFreq}
\title{Reestimate recomb block frequencies}
\usage{
recombFreq(Resp, Block)
}
\arguments{
\item{Resp}{Numerical with the responsibilities for recomb population}

\item{Block}{Character with the blocks genotypes}
}
\value{
Numerical with the new block frequencies for recomb population
}
\description{
Reestimate recomb block frequencies
}
