% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeMatrixColsandRows}
\alias{removeMatrixColsandRows}
\title{Remove columns and Rows from squared matrix}
\usage{
removeMatrixColsandRows(Mat, vIndex)
}
\arguments{
\item{Mat}{Matrix to remove index columns and rows}

\item{vIndex}{Rows and column numbers to be removed from a matrix

<output>}
}
\value{
The original matrix without rows and column index .
}
\description{
removeMatrixColsandRows removes the rows and columns indicated in vector indices vIndex from a matrix
}
