% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setHdf5Group}
\alias{setHdf5Group}
\title{Create a group}
\usage{
setHdf5Group(filename, element, overwrite)
}
\arguments{
\item{filename}{string, path and file name to search element.}

\item{element}{string, group name to be created inside the hdf5 data file.}

\item{overwrite}{boolean, if true this function overwrites existing group.}
}
\value{
integer, if 0 the process was successful and group was created in 
the file
}
\description{
Create a group inside Hdf5 data file
}
