% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_dimNames}
\alias{write_dimNames}
\title{Get dataset dimensions}
\usage{
write_dimNames(filename, group, dataset, rownames, colnames, force = FALSE)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{Character array, indicating the group where the data set is stored}

\item{rownames}{character vector with row names.}

\item{colnames}{character vector with column names.}

\item{datasets, }{Character array, indicating the datasets name column and/or row names belong}
}
\value{
no data returned.
}
\description{
get dataset dimensions
}
