% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CgetData.R
\name{CgetData}
\alias{CgetData}
\title{This function takes a file as input, converts file to gds format
and get data ready to be processed}
\usage{
CgetData(filename, minmaf = 0.1)
}
\arguments{
\item{filename}{character string with file path}

\item{minmaf}{minimum MAF, default 0.01}
}
\value{
data ready to be processed
}
\description{
This function takes a file as input, converts file to gds format
and get data ready to be processed
}
