% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRecombProbabilities_inFile.R
\name{computeRecombProbabilities_inFile}
\alias{computeRecombProbabilities_inFile}
\title{Main function to run all the recombClust pipeline with data in a file}
\usage{
computeRecombProbabilities_inFile(
  filename,
  range,
  samples = NULL,
  window = 500,
  minmaf = 0.1,
  ...
)
}
\arguments{
\item{filename}{File with genotypic data (SNPs and samples) the data file type must be .vcf (Variant Call Format),
.gds (Genomic Data Structure) or .bed (Browser Extensible Data).}

\item{range}{Region GenomicRanges}

\item{window}{only to debug, by default window = 500bp}

\item{...}{Further arguments passed to runLDmixtureModel}

\item{annot}{GenomicRanges with the SNPs annotation}

\item{@param}{samples optional, character vector with sample ids to be used to compute models}
}
\value{
probTab: Summarized probabilities by window
}
\description{
This function takes as input phased chromosomes and returns the responsibilities
PCA and the cluster classification of the samples.
}
