% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createEmptyHdf5File}
\alias{createEmptyHdf5File}
\title{Create File}
\usage{
createEmptyHdf5File(filename, overwrite = FALSE)
}
\arguments{
\item{filename}{string, path and file name to search element.}

\item{overwrite}{boolean, if true this function overwrites existing group.}

\item{element}{string, group name to be created inside the hdf5 data file.}
}
\value{
integer, if 0 the process was successful and group was created in 
the file
}
\description{
Create a group inside Hdf5 data file
}
