% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAnnotationDataHdf5}
\alias{getAnnotationDataHdf5}
\title{Get annotation data from hdf5}
\usage{
getAnnotationDataHdf5(resfilename, resgroup, datasets)
}
\arguments{
\item{resfilename}{string, path and file name where models data is store.}

\item{resgroup}{Annotation obtained for the different models in LDMixture}

\item{datasets}{Range to look in}
}
\value{
Annotation data from models
}
\description{
Get model annotation data from hdf5 data file
}
